// web/projects/portal/src/app/vsobjects/model/layout/zoom-options.ts
/*
 * This file is part of StyleBI.
 * Copyright (C) 2024  InetSoft Technology
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
export enum ZoomOptions {
   ZOOM_IN = -2,
   ZOOM_OUT = -1,

   ZOOM_20 = 0.2,
   ZOOM_40 = 0.4,
   ZOOM_50 = 0.5,
   ZOOM_60 = 0.6,
   ZOOM_75 = 0.75,
   ZOOM_100 = 1,
   ZOOM_125 = 1.25,
   ZOOM_140 = 1.4,
   ZOOM_150 = 1.5,
   ZOOM_160 = 1.6,
   ZOOM_175 = 1.75,
   ZOOM_180 = 1.8,
   ZOOM_200 = 2
}